Refactor: 뒤로가기 버튼 컴포넌트 분리
